Cleaning main.js Moved all db schemcas to model folder
