+++ index.js
@@ -13,8 +13,8 @@ app.use('/static', express.static(path.join(__dirname, 'public')))
 
 //Mongodb connection new 10-22-20
 var mongoose = require('mongoose')
-//var mongoDB ='mongodb://localhost:27017/Inventory'
-var mongoDB = 'mongodb+srv://admin:Pergatory_1979@cluster0.3duu7.mongodb.net/local_library?retryWrites=true&w=majority'
+var mongoDB ='mongodb://localhost:27017/Inventory'
+//var mongoDB = 'mongodb+srv://admin:Pergatory_1979@cluster0.3duu7.mongodb.net/local_library?retryWrites=true&w=majority'
 mongoose.connect(mongoDB,{useNewUrlParser: true, useUnifiedTopology: true});
 var db = mongoose.connection;
 db.on('error', console.error.bind(console,'MongoDB connection error:'));
@@ -57,129 +57,16 @@ app.use(function(req, res, next) {
     next()
 })
 
-////////////////////////////////////// MongoDB Schema Code //////////////////////////////////////
+////////////////////////////////////// MongoDB Schema Routes //////////////////////////////////////
 
-//Database Model for our Cards
-var boardSchema = mongoose.Schema({
-    partNumber: String,
-    serialNumber: String,
-    binNumber: String,
-    binLocation: String,
-    remanPrice: String,
-    exchangePrice: String,
-    dateReceived: String
-});
-var Card = mongoose.model("Card", boardSchema);
-
-//DATABASE MODEL FOR REMOVED CARDS
-var removedboardSchema = mongoose.Schema({
-    partNumber: String,
-    serialNumber: String,
-    binNumber: String,
-    binLocation: String,
-    remanPrice: String,
-    exchangePrice: String,
-    dateRemoved: String
-});
-var RemovedCard = mongoose.model("RemovedCard",removedboardSchema);
-
-//DATABASE MODEL FOR REMOVED PARTS
-var removedPartSchema = mongoose.Schema({
-    stockedAS: String,
-    description1: String,
-    sapNumber: String,
-    manufacturer: String,
-    description2: String,
-    location1: String,
-    location2: String,
-    location3: String,
-    drawer: String,
-    cross1: String,
-    cross2: String,
-    cross3: String,
-    price: String,
-    dateRemoved: String
-});
-var RemovedPart = mongoose.model("RemovedPart",removedPartSchema);
-
-
-//Database Model for Quote requests
-var requestQuoteSchema = mongoose.Schema({
-    partNumber: String,
-    serialNumber: String,
-    remanPrice: String,
-    exchangePrice: String,
-    stockedAS: String,
-    description1: String,
-    sapNumber: String,
-    price: String,
-    name: String,
-    company: String,
-    email: String,
-    phone: String,
-    message: String,
-    dateRequest: String,
-    quantity: String
-});
-var RequestQuote = mongoose.model("RequestQuote",requestQuoteSchema);
-
-//Database Model for our parts
-var partSchema = mongoose.Schema({
-    stockedAS: String,
-    description1: String,
-    sapNumber: String,
-    manufacturer: String,
-    description2: String,
-    location1: String,
-    location2: String,
-    location3: String,
-    drawer: String,
-    cross1: String,
-    cross2: String,
-    cross3: String,
-    price: String
-});
-var Part = mongoose.model("Part", partSchema);
-
-//Database Model for parts RESTOCK
-var restockPartSchema = mongoose.Schema({
-    stockedAS: String,
-    description1: String,
-    sapNumber: String,
-    quantity: String,
-    requestor: String,
-    daterequested: String
-});
-var RestockPart = mongoose.model("RestockPart", restockPartSchema);
-
-//Database Model for Admin Users
-const UserSchema = new mongoose.Schema({
-    nameFirst: {
-      type: String,
-      required: true
-    },
-    nameLast: {
-        type: String,
-        required: true
-      },
-    email: {
-      type: String,
-      required: true
-    },
-    password: {
-      type: String,
-      required: true
-    },
-    token: {
-        type: String
-    },
-    date: {
-      type: Date,
-      default: Date.now
-    }
-  })
-  
-  const User = mongoose.model('User', UserSchema)
+// Load DB Models
+const User = require('./models/User')
+const Card = require('./models/Card')
+const RemovedCard = require('./models/RemovedCard')
+const RemovedPart = require('./models/RemovedPart')
+const RequestQuote = require('./models/RequestQuote')
+const Part = require('./models/Part')
+const RestockPart = require('./models/RestockPart')
 
 ////////////////////////////////////// Cards Route Code //////////////////////////////////////
