+++ models/Template.js
@@ -1,0 +1,5 @@
+const mongoose = require('mongoose')
+
+
+
+module.exports = name
